backend/api_server/requirements.txt
Authlib==0.14.3
connexion[swagger-ui]==2.13.0
Flask-Cors>=3.0.6
gunicorn[gevent] >=20.1.0, <21.0.0
numba # required for where's my gene
numpy==1.21.2  # required for where's my gene
pandas==1.4.3 # required for where's my gene
psutil>=5.9.0
psycopg2-binary>=2.8.5
pyarrow>=1.0 # required for where's my gene
pydantic>=1.9.0
python-jose[cryptography]>=3.1.0
python-json-logger
rsa>=4.7 # not directly required, pinned by Snyk to avoid a vulnerability
scanpy
SQLAlchemy-Utils>=0.36.8
SQLAlchemy>=1.3.17,<2
ddtrace